requirements.txt
openshift-client
prometheus_client
environs
